Add files via upload
finally able to get this to work. This code is beefy as hell, and I'm only just now looking into optimizing it. It needs it desperately. It does seem to work but I do need to confirm that the model is crunching all the number correctly. I would also like to add expanded versions capable of doing hebbian learning, saving to eeprom, and even running a mini simulation with silly emotes and such. Later goals. As of right now it functions slowly but accurately on an arduboy!