Merged matrixToNeuron() into activationFunction()
The compression algorithm made things lag like crazy so we had to find
a way to get things to be quicker without changing too much.
Merging the two of them saved a ton of frames and makes the software
lag significantly less. I added the updated activation function to
both ino files but the matrix version of the code still needs a lot
of updates from the sim, as it is still behind.