Add incomplete sized integer array
Functional, but missing features. Does not support signed numbers.
The underlying array is currently big-endian. It might be preferable to
make it little-endian to match the endianness of avr-gcc. However, since
the ATmega32u4 registers do not have optimized instructions for
arithmetic or bitwise operations on words or larger, either endianness
should be possible to optimize to the same speed.