Matrix up to date
added proper todos to the top, created the options screen. still need to
fix up the serial stuff there a bit, but otherwise its much nicer. Lots
of functionality from the updates in sim were added, as well, such as
faces and support in the diag tool for the gap junctions. Lots still
requires testing though, but we have to wait on the compressed
eeprom for that.