Compressed new data files
The recently updated rom files were compressed using the tool
and saved to the compressed-datasets folder. Though, I still
need to fix the activationFunction() logic in each app for
gap junctionsdata, as this still does not work. Ran into a
snag with a bad upload bricking my arduboy, but will get back
to it soon.